refactors PoolInfoBox into two seperate components and using grid layout to reduce own component clutter.
