+++ packages/ui/components/shared/MidasBox.tsx
@@ -2,12 +2,14 @@ import { Box, BoxProps } from '@chakra-ui/react';
 
 import { useColors } from '@ui/hooks/useColors';
 
-export const PoolDashboardBox = ({ children, ...props }: BoxProps) => {
+export type MidasBoxProps = BoxProps;
+
+export const MidasBox = ({ children, ...props }: MidasBoxProps) => {
   const { cCard } = useColors();
   return (
     <Box
       backgroundColor={cCard.bgColor}
-      borderRadius={10}
+      borderRadius={12}
       borderWidth={2}
       borderColor={cCard.borderColor}
       color={cCard.txtColor}
